+++ src/Controller/Admin/AdminController.php
@@ -66,4 +66,17 @@ final class AdminController extends AbstractController
 
         return new Response(phpinfo());
     }
+
+    /**
+     * @Route("/admin/apcu-cache/clear", name="apcu_cache_clear", methods="GET")
+     * @IsGranted("ROLE_SUPER_ADMIN")
+     */
+    public function clearApcuCache(): Response
+    {
+        apcu_clear_cache();
+
+        $this->addFlash('success', 'APCU Cache was cleared.');
+
+        return $this->redirectToRoute('home');
+    }
 }